requirements.txt
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.2.0
backcall==0.2.0
blinker==1.8.2
ccimport==0.4.4
certifi==2025.1.31
chamferdist==1.0.3
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.1.1
cumm==0.4.11
cumm-cu120==0.4.11
cycler==0.12.1
dash==3.1.0
decorator==5.2.1
diffusers==0.32.2
docker-pycreds==0.4.0
e3nn==0.5.6
einops==0.8.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.7.0
flash-attn==2.7.0.post2
Flask==3.0.3
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.29.3
hydra-core==1.3.2
idna==3.10
imageio==2.31.6
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
ipython==8.12.3
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.9
lark==1.2.2
lightning==2.2.2
lightning-utilities==0.11.9
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
narwhals==1.42.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
ocnn @ git+https://github.com/octree-nn/ocnn-pytorch.git@a0e2717427f79efaee82a680c0de7c445c7e3fb6
omegaconf==2.3.0
open3d==0.19.0
opt-einsum-fx==0.1.4
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pathtools==0.1.2
pccm==0.4.16
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.2.0
pointops @ file:///home/cederic/dev/puzzlefusion-plusplus/Pointcept/libs/pointops
portalocker==3.0.0
prompt_toolkit==3.0.51
propcache==0.2.0
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==2.13.6
Pygments==2.19.2
pyparsing==3.1.4
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
-e git+https://github.com/facebookresearch/pytorch3d.git@21205730d9214768d9076e48147433818f0bb58e#egg=pytorch3d
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.4.0
rpds-py==0.20.1
Rtree==1.3.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==2.22.0
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
spconv-cu120==2.3.6
stack-data==0.6.3
sympy==1.13.3
termcolor==2.4.0
threadpoolctl==3.5.0
timm==1.0.15
tokenizers==0.20.3
torch==2.4.1
torch-geometric==2.6.1
torch-summary==1.4.5
torch_cluster==1.6.3
torch_scatter==2.1.2
torch_sparse==0.6.18
torchaudio==2.4.1
torchinfo==1.8.0
torchmetrics==1.5.2
torchsort==0.1.10
torchvision==0.19.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
trimesh==4.0.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
wandb==0.15.12
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.14
yarl==1.15.2
zipp==3.20.2
